Example data for deweather

The dat dataset is provided as an example dataset as part of the deweather package. The dataset contains hourly measurements of air pollutant concentrations, wind speed and wind direction collected at the Marylebone (London) air quality monitoring supersite between 1st January 1998 and 23rd June 2005.

Usage

dat

Format

Data frame with 65533 observations (rows) on the following 10 variables:

list("date")

Observation date/time stamp in year-month-day hour:minute:second format (POSIXct).

list("ws")

Wind speed, in m/s, as numeric vector.

list("wd")

Wind direction, in degrees from North, as a numeric vector.

list("nox")

Oxides of nitrogen concentration, in ppb, as a numeric vector.

list("no2")

Nitrogen dioxide concentration, in ppb, as a numeric vector.

list("o3")

Ozone concentration, in ppb, as a numeric vector.

list("pm10")

Particulate PM10 fraction measurement, in ug/m3 (raw TEOM), as a numeric vector.

list("so2")

Sulfur dioxide concentration, in ppb, as a numeric vector.

list("co")

Carbon monoxide concentration, in ppm, as a numeric vector.

list("pm25")

Particulate PM2.5 fraction measurement, in ug/m3, as a numeric vector.

Source

mydata was compiled from data archived in the London Air Quality Archive. See http://www.londonair.org.uk for site details.

The same data is also provide in '.csv' format via the openair project web site http://www.openair-project.org.

Details

dat is supplied with the rmweather package as an example dataset for use with documented examples.

Note

openair functions generally require data frames with a field "date" that can be in either POSIXct or Date format but should be GMT time zone. This can be hourly data or higher resolution data.

Examples

# basic structure
head(dat)
#>                  date ethane nox no2 o3       ws  wd cl visibility pressure
#> 1 2012-01-01 00:00:00   6.09 157  67 22 6.687778 220  8       3500   1008.1
#> 2 2012-01-01 01:00:00   6.71 310  96 12 7.716667 240  8       2900   1008.0
#> 3 2012-01-01 02:00:00   6.73 233  90 16 6.173333 220  5       1500   1008.0
#> 4 2012-01-01 03:00:00   6.49 166  59 20 6.173333 220  5       4000   1007.7
#> 5 2012-01-01 04:00:00   6.69 147  59 26 4.630000 220  4       3500   1007.6
#> 6 2012-01-01 05:00:00   6.99 132  52 26 5.658889 220  6       4000   1007.5
#>   temp   rh rain
#> 1 12.2 90.3  0.2
#> 2 12.4 89.2  0.0
#> 3 11.9 93.8  0.0
#> 4 11.8 92.6  0.0
#> 5 11.7 91.4  0.0
#> 6 11.9 90.2  0.0