deweather: Remove the influence of weather on air quality data

Model and account for (or remove) the effect of meteorology on atmospheric composition data. The technique uses boosted regression trees via the tidymodels framework.

See also

Author

Maintainer: David Carslaw david.carslaw@york.ac.uk (ORCID)

Authors:

Jack Davison jack.davison@ricardo.com (ORCID)